Let the filters be appendable
