+++ src/components/Envoy/tables/BaseTable.tsx
@@ -5,7 +5,7 @@ import { RouteSummaryTable, RouteTable } from './RouteTable';
 import { ListenerSummaryTable, ListenerTable } from './ListenerTable';
 import { EnvoyProxyDump } from '../../../types/IstioObjects';
 import { FilterSelected, StatefulFilters } from '../../Filters/StatefulFilters';
-import { FilterType } from '../../../types/Filters';
+import { ActiveFilter, ActiveFiltersInfo, FilterType } from '../../../types/Filters';
 import { setFiltersToURL } from '../../FilterList/FilterHelper';
 import { ResourceSorts } from '../EnvoyModal';
 
@@ -74,4 +74,40 @@ export const SummaryTableBuilder = (resource: string, config: EnvoyProxyDump, so
       break;
   }
   return [writerComp, writerProps];
+};
+export const defaultFilter = (value: (string | number)[], filterToColumn: { [id: string]: number }): boolean => {
+  const activeFilters: ActiveFiltersInfo = FilterSelected.getSelected();
+  // If there is no active filters, show the entry
+  if (activeFilters.filters.length === 0) {
+    return true;
+  }
+
+  // Group filters by id
+  const groupedFilters: ActiveFilter[][] = activeFilters.filters.reduce(
+    (groupedFilters: ActiveFilter[][], filter: ActiveFilter): ActiveFilter[][] => {
+      let filterGroup = groupedFilters[filterToColumn[filter.id]];
+      if (!filterGroup) {
+        filterGroup = [];
+      }
+      groupedFilters[filterToColumn[filter.id]] = filterGroup.concat(filter);
+      return groupedFilters;
+    },
+    []
+  );
+
+  // Show entities that has a match in each filter group
+  return groupedFilters.reduce((prevMatch: boolean, filters: ActiveFilter[]): boolean => {
+    // There is at least one filter matching the item in the group
+    return (
+      prevMatch &&
+      filters.some((filter: ActiveFilter) => {
+        const row: number = filterToColumn[filter.id];
+        let match: boolean = true;
+        if (row !== undefined) {
+          match = value[row].toString().includes(filter.value);
+        }
+        return match;
+      })
+    );
+  }, true);
 };
